refactor(spring-redis): update connection config types and API usage
- Change exponent_base type from u64 to f32 in RedisConfig struct
- Remove unnecessary cast when passing exponent_base to set_exponent_base()
- Wrap max_delay in Duration::from_millis() for proper type conversion
- Wrap response_timeout Duration in Some() to match updated API signature
- Wrap connection_timeout Duration in Some() to match updated API signature
- Align with upstream redis connection manager API changes for timeout handling

+++ spring-redis/src/config.rs
@@ -22,7 +22,7 @@ pub struct RedisConfig {
 
     /// The resulting duration is calculated by taking the base to the `n`-th power,
     /// where `n` denotes the number of past attempts.
-    pub exponent_base: Option<u64>,
+    pub exponent_base: Option<f32>,
 
     /// Apply a maximum delay between connection attempts. The delay between attempts won't be longer than max_delay milliseconds.
     pub max_delay: Option<u64>,
